Extract aztec click listener
