+++ app/src/main/java/ar/edu/unq/pdes/myprivateblog/screens/post_create/PostCreateFragment.kt
@@ -9,12 +9,8 @@ import androidx.navigation.fragment.findNavController
 import ar.edu.unq.pdes.myprivateblog.BaseFragment
 import ar.edu.unq.pdes.myprivateblog.ColorUtils
 import ar.edu.unq.pdes.myprivateblog.R
+import ar.edu.unq.pdes.myprivateblog.utils.setAztecToolbarClickListener
 import kotlinx.android.synthetic.main.fragment_post_edit.*
-import org.wordpress.aztec.Aztec
-import org.wordpress.aztec.ITextFormat
-import org.wordpress.aztec.glideloader.GlideImageLoader
-import org.wordpress.aztec.glideloader.GlideVideoThumbnailLoader
-import org.wordpress.aztec.toolbar.IAztecToolbarClickListener
 
 class PostCreateFragment : BaseFragment(R.layout.fragment_post_edit) {
     private val viewModel by viewModels<PostCreateViewModel> { viewModelFactory }
@@ -32,7 +28,7 @@ class PostCreateFragment : BaseFragment(R.layout.fragment_post_edit) {
                         PostCreateFragmentDirections.navActionSaveNewPost(viewModel.post)
                     )
                 }
-                else -> { /* Do nothing, should not happen*/
+                else -> {
                 }
             }
         })
@@ -60,22 +56,6 @@ class PostCreateFragment : BaseFragment(R.layout.fragment_post_edit) {
             viewModel.cardColor.postValue(it)
         }
 
-        context?.apply {
-            Aztec.with(body, source, formatting_toolbar, object : IAztecToolbarClickListener {
-                override fun onToolbarCollapseButtonClicked() {}
-                override fun onToolbarExpandButtonClicked() {}
-                override fun onToolbarHeadingButtonClicked() {}
-                override fun onToolbarHtmlButtonClicked() {}
-                override fun onToolbarListButtonClicked() {}
-                override fun onToolbarMediaButtonClicked(): Boolean = false
-                override fun onToolbarFormatButtonClicked(
-                    format: ITextFormat,
-                    isKeyboardShortcut: Boolean
-                ) {
-                }
-            })
-                .setImageGetter(GlideImageLoader(this))
-                .setVideoThumbnailGetter(GlideVideoThumbnailLoader(this))
-        }
+        context?.apply { setAztecToolbarClickListener(this, body, source, formatting_toolbar) }
     }
 }
